# Force specialization on the type argument of `_similar_for`