Looking into the `["misc", "iterators", "zip(1:1, 1:1, 1:1, 1:1)"] ` regression reported in https://github.com/JuliaLang/julia/pull/30218#issuecomment-443746637, it looks like #28284 pushed the `HasShape` method of `_similar_for` beyond the inlining threshold for that case, causing the regression. Adding a `@_meta_inline` would directly combat that, but forcing specialization on the type argument is less invasive and still restores performance:
```julia
julia> for N in (1,1000), M in (1,2,3,4)
           @btime collect(zip($(Iterators.repeated(1:N, M)...,)...))
       end
# master
  30.213 ns (1 allocation: 96 bytes)
  32.706 ns (1 allocation: 96 bytes)
  38.512 ns (1 allocation: 112 bytes)
  130.325 ns (5 allocations: 272 bytes)
  1.069 μs (1 allocation: 7.94 KiB)
  1.753 μs (1 allocation: 15.75 KiB)
  2.367 μs (2 allocations: 23.52 KiB)
  3.187 μs (6 allocations: 31.48 KiB)
# this PR
  30.425 ns (1 allocation: 96 bytes)
  32.565 ns (1 allocation: 96 bytes)
  37.667 ns (1 allocation: 112 bytes)
  42.188 ns (1 allocation: 112 bytes)
  1.048 μs (1 allocation: 7.94 KiB)
  1.781 μs (1 allocation: 15.75 KiB)
  2.359 μs (2 allocations: 23.52 KiB)
  2.837 μs (2 allocations: 31.33 KiB)
```

I hope @nanosoldier `runbenchmarks(ALL, vs = ":master")` agrees. In that case, I guess we want this on 1.1 as well.

I vaguely remember having noticed this when working on #29238 but decided it was an orthogonal change to be done in a separate PR and then forgot about it...