@nanosoldier 2018-12-10
[Your benchmark job](https://github.com/JuliaLang/julia/pull/30331) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/352a94d4d60da5a44aba53eaeccd3e3376118b79/b6430e2_vs_0a401f2/report.md). cc @ararslan