@martinholters 2018-12-11
OSX Travis ended with:
```
1788.795694  cd /Users/travis/build/JuliaLang/julia/base && if ! DYLD_FALLBACK_LIBRARY_PATH="/usr/local/lib:/lib:/usr/lib:/usr/local/opt/openblas-julia/lib:/usr/local/opt/suite-sparse-julia/lib" /Users/travis/build/JuliaLang/julia/usr/bin/julia-debug -O0 -C "generic;native" --output-o /Users/travis/build/JuliaLang/julia/usr/lib/julia/sys-debug-o.a.tmp  --startup-file=no --warn-overwrite=yes --sysimage /Users/travis/build/JuliaLang/julia/usr/lib/julia/sys.ji /Users/travis/build/JuliaLang/julia/contrib/generate_precompile.jl 1; then echo '*** This error is usually fixed by running `make clean`. If the error persists, try `make cleanall`. ***'; false; fi 
No output has been received in the last 30m0s, this potentially indicates a stalled build or something wrong with the build itself.
Check the details on how to adjust your build configuration on: https://docs.travis-ci.com/user/common-build-problems/#Build-times-out-because-no-output-was-received
The build has been terminated
```
Restarted.

Benchmarks look good, but were run before I also changed the `AbstractSet` / `AbstractDict` methods, so better safe than sorry: @nanosoldier `runbenchmarks(ALL, vs = ":master")`