@nanosoldier 2018-12-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/30331#issuecomment-446103696) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/3d9bd49f8af21ef8b709ee67e48597abbfd23fad/b2f94cb_vs_0d62000/report.md). cc @ararslan