@martinholters 2018-12-12
What's our policy for backporting performance enhancements to 1.0?