@StefanKarpinski 2018-12-12
If they're low risk then absolutely do it.